remove Input file related stuff from the compiler and runtime system

+++ Robust/src/ClassLibrary/MGC/gnu/Float.java
@@ -355,7 +355,7 @@ public final class Float
   public static float parseFloat(String str)
   {
     //return VMFloat.parseFloat(str);
-    return (float)(Long.parseLong(str));
+    return (float)(Double.parseDouble(str));
   }
 
   /**
